Init Commit.
Currently implemented: Read Method.
Returns a boolean on if the read was successful, if true, out variable will contain an array of movies obtained from the CSV file. Otherwise it will be empty.

Write method still needs to be implemented. Additionally some form of pretty print for movies should be created.